gemi (voice): fix model path and unzipping script

+++ install_models.sh
@@ -1,15 +1,16 @@
 #!/bin/bash
 
 for t in `curl http://cmuflite.org/packed/flite-2.3/voices/us_voices`; do
-    curl -o voices/flite/$t http://cmuflite.org/packed/flite-2.3/voices/$t
+    if [ ! -f voices/flite/$t ]; then
+        curl -o voices/flite/$t http://cmuflite.org/packed/flite-2.3/voices/$t
+    fi
 done
 
 mkdir tmp
-cd tmp
+curl -o tmp/vosk-model-en-in-0.5.zip https://alphacephei.com/vosk/models/vosk-model-en-in-0.5.zip
 
-curl -O https://alphacephei.com/vosk/models/vosk-model-en-in-0.5.zip
-unzip -o vosk-model-en-in-0.5.zip
-mv vosk-model-en-in-0.5 models/vosk-model-en-in-0.5
+cd models
+unzip -o ../tmp/vosk-model-en-in-0.5.zip
 
 cd ..
 rm -rf tmp
